+++ src/main/java/com/hanwol/service/StageService.java
@@ -25,7 +25,7 @@ public class StageService {
      * 특정 막(Chapter)의 스테이지 목록을 유저 진행도와 함께 조회
      */
     @Transactional(readOnly = true)
-    public List<StageResponse> getStagesByChapter(Long userId, Integer chapterId) {
+    public List<StageResponse> getStagesByChapter(Long userId, Long chapterId) {
         UserProgress progress = userProgressRepository.findById(userId)
                 .orElseGet(() -> createInitialProgress(userId));
 
@@ -35,7 +35,7 @@ public class StageService {
             boolean isCleared = stage.getId() <= progress.getMaxClearedStageId();
             // 이전 스테이지를 깼거나, 첫 스테이지면 해금
             boolean isLocked = stage.getId() > (progress.getMaxClearedStageId() + 1);
-            
+
             return new StageResponse(stage, isCleared, isLocked);
         }).collect(Collectors.toList());
     }
@@ -44,7 +44,7 @@ public class StageService {
      * 스테이지 진입 전 스토리 정보 확인
      */
     @Transactional(readOnly = true)
-    public Optional<Integer> getBeforeStoryId(Integer stageId, Long userId) {
+    public Optional<Long> getBeforeStoryId(Long stageId, Long userId) {
         // 이미 클리어한 스테이지면 스토리를 보여주지 않음 (도감에서 보게 유도)
         UserProgress progress = userProgressRepository.findById(userId).orElse(null);
         if (progress != null && stageId <= progress.getMaxClearedStageId()) {
@@ -67,5 +67,6 @@ public class StageService {
     }
 
     // 내부 응답용 DTO
-    public record StageResponse(Stage stage, boolean isCleared, boolean isLocked) {}
+    public record StageResponse(Stage stage, boolean isCleared, boolean isLocked) {
+    }
 }
